customize(router): move Websocket setup after session validation
We do not want to start Websocket communication until the session was
validated. Background: we want to avoid as much OpenWebUI traffic ac
possible.

I do not know the code well enough _yet_ to understand whether the
Websocket was intentionally started before validating the session
and I don't want to invest time in analyzing it right now, so this
might break things subtly.