fix(UserMenu): with OIDC logout don't follow redirect in fetch
== Background

When OIDC is configured, the session in Open WebUI is invalidated and
the user must be ridirected to the auth server's end_session_endpoint
to invalidate the session there.

== Problem

With the previous implementation, the end_session_endpoint was sent as a
redirect to the client.

However, since the request is made as Fetch request it has to obey CORS
rules.

This would require the Auth server to set Access-Control-Allow-Origin,
which is not sensible. If this header is not set, the request would fail,
thus the logout would not be completed at the Auth server.

== Fix

Instead of sending a redirect the target URL is sent to the client,
which then can be used to change the user agent's location.

== Discarded approach

The Fetch option `redirect: manual` [1] was not chosen because it would
not change the user agent's location to the OIDC provider's logout page.

With this approach the end_session_endpoint URL would be presented to
the user.

[1]: https://developer.mozilla.org/en-US/docs/Web/API/RequestInit#redirect