c/fix(github/workflows): use x-access-token + GITHUB_TOKEN
Before the previous commit the Job creation failed with unknown reason.
The previous commit added more logging and revealed the clone failed:

   Submodule 'backend/custom_resources' (https://github.com/IONOS-Productivity/gpt-open-webui-custom-resources) registered for path 'backend/custom_resources'
   Cloning into '/workspace/backend/custom_resources'...
   remote: Invalid username or password.
   fatal: Authentication failed for 'https://github.com/IONOS-Productivity/gpt-open-webui-custom-resources/'

Although I expected the "oauth:xxx" to be shown I hazard a guess this
was just a logging artifact by GitHub actions and the Job actually
failed due to an invalid token.