c/feat(services/auths): explicitly configure logout endpoint URI for OIDC logout too
The post-logout URI was specified only for the non-OIDC-compliant logout
URL.

However it's better to explicitly pass it to the OIDC
end_session_endpoint too to have the same behavior when testing with
Minikube.

This prepares using the "post_logout_redirect_uri" or "redirect_url" for
as destination after logout for password reset.

Refs: PRODAI-90