c/fix(Scroller): fix non-auto-starting scroller
== Cause

The scroller animation did not start because the `animation-duration`
property had the value 0 initially.

This was caused by `el.scrollWidth` being 0 initially because there
were no elements in `items`, which was caused by PromptSelector loading
the items asynchroneously on mount.

== Fix

The reactive expression now incorporates the state of `items` for
recalculation.