[WIP] c/chore(Dockerfile): build agents json (static) during build
== Background

static/avatars/ and static/data are agent avatar images and agent
configs as static JSON. Their source data is part of the
custom_resources project.

These JSON files and images have to be generated in the root project to
make them available to the SPA.

== Change

This change generates them during image build.

Refs: PRODAI-329