c/refac(KnowledgeManager): always render edit & create dialog
And show them by "show" prop.

== Previously

Overlays were only rendered with "show" = true depending on their
state.

For the EditKnowledge dialog this was necessary because it requires
"knowledgeBeingEdited" to be defined.

== Now

The overlays are always rendered and their state depends on the
"show" prop.

EditKnowledge is now adapted to handle the "knowledge" prop to be null.

== Rationale

Prepare adapting the dialogs for mobile where they have to slide in/out.
Therefore they have to stay in DOM for the transition to start and end
on show state.

Refs: PRODAI-335