c/fix(Dialog): attempt to fix collapsed dialog in Safari 17
== Problem

In Safari 17.3 on iOS 15 the dialog is collapsed to a couple of pixels.

Narrowing down the issue revealed that, if the first-child div in the dialog
element has a height: 100% this error is triggered:

   <dialog>
     <div style="height: 100%;">
       <p>lorem ipsum dolor sit amet lorem ipsum dolor sit amet lorem ipsum dolor sit amet lorem ipsum dolor sit amet lorem ipsum dolor sit amet lorem ipsum dolor sit amet lorem ipsum dolor sit amet </p>
       <p>lorem ipsum dolor sit amet lorem ipsum dolor sit amet lorem ipsum dolor sit amet lorem ipsum dolor sit amet lorem ipsum dolor sit amet lorem ipsum dolor sit amet lorem ipsum dolor sit amet </p>
     </div>
   </dialog>

== Solution

It turns out that removing the height fixes the problem for Safari.

However, it is needed for Chrome to consume all height for the flex
container.

Using min-height: 100% fixes the problem for Safai, while retaining
functionality for Chrome.

Refs: SMBMOCQA-8550