fix(requirements.txt): set posthog version < 6.0
== Error

The application would log the error message

   ERROR [chromadb.telemetry.product.posthog] Failed to send telemetry event ClientStartEvent: capture() takes 1 positional argument but 3 were given

on startup. This is due to the newest version of posthog (6.x)
not being API compatible with the installed ChromaDB version.

== Cause

This is due to the Dockerfile always installing versions according to
requirements.txt, uv.lock is not regarded.

== Note

Coincidentally dicovered while changing the Dockerfile (while a cached
layer could no longer be used).

== See also

* https://github.com/chroma-core/chroma/issues/4966
* https://github.com/open-webui/open-webui/issues/15673

Refs: PRODAI-421