Fix broken i18n access
This fixes an issue where the translations do not always show up for the
agent details.

The when you called agentCardData() on initial render it might access
the i18n store before it had been initialized. It would then return
the default string (key name) as replacement, not the text.